chore: update code structure and remove redundant sections

+++ routes/web.php
@@ -10,6 +10,7 @@ use App\Http\Controllers\MapDataController;
 use App\Http\Controllers\ReportController;
 use App\Http\Controllers\UniversityController;
 use Illuminate\Support\Facades\App;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 
 Route::redirect('/', '/es', 301);
@@ -17,6 +18,9 @@ Route::redirect('/login', '/', 301);
 
 Route::group(['prefix' => '{locale}', 'where' => ['locale' => 'es|en']], function () {
     Route::get('/', function (string $locale) {
+        if (Auth::check()) {
+            return redirect(route('events'));
+        }
         App::setLocale($locale);
         return view('index', ['locale' => $locale]);
     })->name('home');
@@ -37,11 +41,11 @@ Route::post('/{locale}/assistance/verify', [AssistanceController::class, 'verify
 Route::prefix('dashboard')->middleware(['auth', 'verified'])->group(function () {
     Route::get('/', function () {
         return view('dashboard.index');
-    })->name('dashboard');
+    })->name('dashboard')->middleware('role:admin');
 
     Route::prefix('regen')->group(function () {
         Route::get('/', function () {
-            return view('dashboard.pages.regen.index');
+            return redirect()->route('events');
         })->name('dashboard.regen');
 
         Route::get('/signatures', function () {
@@ -53,49 +57,51 @@ Route::prefix('dashboard')->middleware(['auth', 'verified'])->group(function () 
     Route::prefix('events')->group(function () {
 
         Route::get('/', [EventController::class, 'index'])->name('events');
-        Route::post('/generate-report', [ReportController::class, 'generateReport'])->name('generate.report');
+        Route::post('/generate-report', [ReportController::class, 'generateReport'])->name('generate.report')->middleware('role:admin');
         Route::get('/edit/{id}', [EventController::class, 'edit'])->name('events.edit');
-        Route::post('/', [EventController::class, 'store'])->name('events.store');
-        Route::put('/{id}', [EventController::class, 'update'])->name('events.update');
-        Route::delete('/{id}', [EventController::class, 'destroy'])->name('events.destroy');
-        Route::post('/events/{id}/certificates', [CertificateController::class, 'sendAllCertificates'])->name('events.sendAllCertificates');
-        Route::post('/events/{event_id}/certificates/{assistance_id}', [CertificateController::class, 'sendCertificate'])->name('events.sendCertificate');
+
+        Route::post('/', [EventController::class, 'store'])->name('events.store')->middleware('role:admin');
+        Route::put('/{id}', [EventController::class, 'update'])->name('events.update')->middleware('role:admin');
+        Route::delete('/{id}', [EventController::class, 'destroy'])->name('events.destroy')->middleware('role:admin');
+        Route::post('/events/{id}/certificates', [CertificateController::class, 'sendAllCertificates'])->name('events.sendAllCertificates')->middleware('role:admin');
+
+        Route::post('/events/{event_id}/certificates/{assistance_id}', [CertificateController::class, 'sendCertificate'])->name('events.sendCertificate')->middleware('role:admin');
 
         Route::post('/events/{event_id}/certificates/aprove', [CertificateController::class, 'approveCertificate'])->name('events.approveCertificate')->middleware('role:regen');
     });
 
     Route::prefix('universities')->group(function () {
-        Route::get('/', [UniversityController::class, 'index'])->name('universities');
-        Route::get('/{id}', [UniversityController::class, 'edit'])->name('universities.edit');
-        Route::post('/', [UniversityController::class, 'store'])->name('universities.store');
-        Route::put('/{id}', [UniversityController::class, 'update'])->name('universities.update');
-        Route::delete('/{id}', [UniversityController::class, 'destroy'])->name('universities.destroy');
+        Route::get('/', [UniversityController::class, 'index'])->name('universities')->middleware('role:admin');
+        Route::get('/{id}', [UniversityController::class, 'edit'])->name('universities.edit')->middleware('role:admin');
+        Route::post('/', [UniversityController::class, 'store'])->name('universities.store')->middleware('role:admin');
+        Route::put('/{id}', [UniversityController::class, 'update'])->name('universities.update')->middleware('role:admin');
+        Route::delete('/{id}', [UniversityController::class, 'destroy'])->name('universities.destroy')->middleware('role:admin');
     });
 
     Route::prefix('activities')->group(function () {
-        Route::get('/', [ActivityController::class, 'index'])->name('activities');
-        Route::get('/{id}', [ActivityController::class, 'edit'])->name('activities.edit');
-        Route::post('/', [ActivityController::class, 'store'])->name('activities.store');
-        Route::put('/{id}', [ActivityController::class, 'update'])->name('activities.update');
-        Route::delete('/{id}', [ActivityController::class, 'destroy'])->name('activities.destroy');
-    });
+        Route::get('/', [ActivityController::class, 'index'])->name('activities')->middleware('role:admin');
+        Route::get('/{id}', [ActivityController::class, 'edit'])->name('activities.edit')->middleware('role:admin');
+        Route::post('/', [ActivityController::class, 'store'])->name('activities.store')->middleware('role:admin');
+        Route::put('/{id}', [ActivityController::class, 'update'])->name('activities.update')->middleware('role:admin');
+        Route::delete('/{id}', [ActivityController::class, 'destroy'])->name('activities.destroy')->middleware('role:admin');
+    })->middleware('role:admin');
 
     Route::prefix('agreements')->group(function () {
-        Route::get('/', [AgreementController::class, 'index'])->name('agreements');
-        Route::get('/{id}', [AgreementController::class, 'show'])->name('agreements.show');
-        Route::post('/', [AgreementController::class, 'store'])->name('agreements.store');
-        Route::put('/{id}', [AgreementController::class, 'update'])->name('agreements.update');
-        Route::get('/{id}', [AgreementController::class, 'edit'])->name('agreements.edit');
-        Route::delete('/{id}', [AgreementController::class, 'destroy'])->name('agreements.destroy');
-    });
+        Route::get('/', [AgreementController::class, 'index'])->name('agreements')->middleware('role:admin');
+        Route::get('/{id}', [AgreementController::class, 'show'])->name('agreements.show')->middleware('role:admin');
+        Route::post('/', [AgreementController::class, 'store'])->name('agreements.store')->middleware('role:admin');
+        Route::put('/{id}', [AgreementController::class, 'update'])->name('agreements.update')->middleware('role:admin');
+        Route::get('/{id}', [AgreementController::class, 'edit'])->name('agreements.edit')->middleware('role:admin');
+        Route::delete('/{id}', [AgreementController::class, 'destroy'])->name('agreements.destroy')->middleware('role:admin');
+    })->middleware('role:admin');
 
     Route::prefix('careers')->group(function () {
-        Route::get('/', [CareerController::class, 'index'])->name('careers');
-        Route::get('/{id}', [CareerController::class, 'edit'])->name('careers.edit');
-        Route::post('/', [CareerController::class, 'store'])->name('careers.store');
-        Route::put('/{id}', [CareerController::class, 'update'])->name('careers.update');
-        Route::delete('/{id}', [CareerController::class, 'destroy'])->name('careers.destroy');
-    });
+        Route::get('/', [CareerController::class, 'index'])->name('careers')->middleware('role:admin');
+        Route::get('/{id}', [CareerController::class, 'edit'])->name('careers.edit')->middleware('role:admin');
+        Route::post('/', [CareerController::class, 'store'])->name('careers.store')->middleware('role:admin');
+        Route::put('/{id}', [CareerController::class, 'update'])->name('careers.update')->middleware('role:admin');
+        Route::delete('/{id}', [CareerController::class, 'destroy'])->name('careers.destroy')->middleware('role:admin');
+    })->middleware('role:admin');
 });
 
 require __DIR__ . '/auth.php';
